+++ entsoe_api/api.py
@@ -1,5 +1,4 @@
 import requests
-import xml.etree.ElementTree as ET
 import pandas as pd
 from datetime import datetime, timedelta
 
@@ -19,10 +18,9 @@ class EntsoeAPI:
     """
 
     BASE_URL = 'https://web-api.tp.entsoe.eu/api'
-    MAX_PERIOD_DAYS = 30
     REQUEST_DELAY = 0.5
 
-    def __init__(self, api_key: str):
+    def __init__(self, api_key: str, max_period_days: int = 30):
         """Initializes the EntsoeAPI with the given API key.
 
         Args:
@@ -30,10 +28,11 @@ class EntsoeAPI:
         """
         self.api_key = api_key
         self.data_parser = DataParser()
+        self.max_period_days = max_period_days
 
     def _get_data(
             self, start_date: datetime, end_date: datetime, document_type: DocumentType,
-            process_type: ProcessType, domain: DomainType, psr_type: PsrType = 'ALL',
+            process_type: ProcessType, in_domain: DomainType, out_domain: DomainType, psr_type: PsrType = 'ALL',
     ) -> bytes:
         """Fetches data from the ENTSO-E API.
 
@@ -44,7 +43,8 @@ class EntsoeAPI:
             end_date (datetime): The end date and time of the data request.
             document_type (DocumentType): The type of document to request.
             process_type (ProcessType): The type of process to request.
-            domain (DomainType): The domain code to specify the area.
+            in_domain (DomainType): The domain code to specify the area.
+            out_domain (DomainType): The domain code to specify the area.
             psr_type (PsrType, optional): The type of generation source. Defaults to 'ALL'.
 
         Returns:
@@ -56,8 +56,9 @@ class EntsoeAPI:
         params = {
             'documentType': document_type.value,
             'processType': process_type.value,
-            'in_Domain': domain.value,
-            'out_Domain': domain.value,
+            'in_Domain': in_domain.value,
+            'out_Domain': out_domain.value,
+            'OutBiddingZone_Domain': in_domain.value,
             'periodStart': start_date.strftime('%Y%m%d%H00'),
             'periodEnd': end_date.strftime('%Y%m%d%H00'),
             'securityToken': self.api_key
@@ -75,20 +76,20 @@ class EntsoeAPI:
 
     def fetch_data(
             self, start_date: datetime, end_date: datetime, document_type: DocumentType,
-            process_type: ProcessType, domain: DomainType, psr_type: PsrType = 'ALL',
+            process_type: ProcessType, in_domain: DomainType, out_domain: DomainType | None = None, psr_type: PsrType = 'ALL',
     ) -> pd.DataFrame:
         """Fetches data from the ENTSO-E API.
 
         This method constructs the request to fetch energy production data between the specified start and end dates.
 
-        Args:
+       Args:
             start_date (datetime): The start date and time of the data request.
             end_date (datetime): The end date and time of the data request.
             document_type (DocumentType): The type of document to request.
             process_type (ProcessType): The type of process to request.
-            domain (DomainType): The domain code to specify the area.
+            in_domain (DomainType): The domain code to specify the area for incoming data.
+            out_domain (DomainType, optional): The domain code to specify the area for outgoing data. Defaults to None.
             psr_type (PsrType, optional): The type of generation source. Defaults to 'ALL'.
-
         Returns:
             pd.DataFrame: A pandas DataFrame containing the production data.
 
@@ -99,28 +100,36 @@ class EntsoeAPI:
 
         delta_days = (end_date - start_date).days
 
-        if delta_days > self.MAX_PERIOD_DAYS:
-            LOGGER.debug(f'Date range exceeds {self.MAX_PERIOD_DAYS} days. Splitting the request.')
+        if delta_days > self.max_period_days:
+            LOGGER.debug(f'Date range exceeds {self.max_period_days} days. Splitting the request.')
 
             # Split the date range into chunks of at most MAX_PERIOD_DAYS days
             data_frames = []
 
-            for i in range(0, delta_days, self.MAX_PERIOD_DAYS):
+            for i in range(0, delta_days, self.max_period_days):
                 chunk_start = start_date + timedelta(days=i)
                 next_step = chunk_start + timedelta(
-                    days=self.MAX_PERIOD_DAYS - 1
+                    days=self.max_period_days - 1
                     if document_type == DocumentType.PRICE_DOCUMENT
-                    else self.MAX_PERIOD_DAYS
+                    else self.max_period_days
                 )
-                chunk_end = min(next_step, end_date)
+                chunk_end = min(next_step, end_date - timedelta(days=1) if document_type == DocumentType.PRICE_DOCUMENT else end_date)
 
                 data_frames.append(
-                    self.fetch_data(chunk_start, chunk_end, document_type, process_type, domain, psr_type))
+                    self.fetch_data(chunk_start, chunk_end, document_type, process_type, in_domain, out_domain, psr_type)
+                )
 
             return pd.concat(data_frames)
         else:
-            LOGGER.debug('Fetching data from ENTSO-E API...')
-            xml_data = self._get_data(start_date, end_date, document_type, process_type, domain, psr_type)
-            LOGGER.debug('Parsing data...')
+            LOGGER.debug(f'Fetching data from ENTSO-E API for {start_date} to {end_date}...')
+            xml_data = self._get_data(
+                start_date=start_date,
+                end_date=end_date,
+                document_type=document_type,
+                process_type=process_type,
+                in_domain=in_domain,
+                out_domain=out_domain if out_domain is not None else in_domain,
+                psr_type=psr_type
+            )
             df = self.data_parser.parse_data(xml_data, document_type)
             return df