+++ entsoe_api/parser/data_parser.py
@@ -1,8 +1,10 @@
 import pandas as pd
 
 from entsoe_api.enums import DocumentType
+from entsoe_api.parser.parser_extensions.aggregated_energy_data_data_parser import AggregatedEnergyDataDataParser
 from entsoe_api.parser.parser_extensions.price_data_parser import PriceDataParser
 from entsoe_api.parser.parser_extensions.production_data_parser import ProductionDataParser
+from entsoe_api.parser.parser_extensions.total_load_data_parser import TotalLoad
 from entsoe_api.parser.parser_interface import ParserInterface
 from entsoe_api.parser.exceptions import ParserError
 
@@ -14,6 +16,9 @@ class DataParser:
         self.parsers: dict[DocumentType, type[ParserInterface]] = {
             DocumentType.ACTUAL_GENERATION_PER_TYPE: ProductionDataParser,
             DocumentType.PRICE_DOCUMENT: PriceDataParser,
+            DocumentType.AGGREGATED_ENERGY_DATA_REPORT: AggregatedEnergyDataDataParser,
+            DocumentType.WIND_AND_SOLAR_FORECAST: ProductionDataParser,
+            DocumentType.SYSTEM_TOTAL_LOAD: TotalLoad,
         }
 
     def parse_data(self, xml_data: bytes, document_type: DocumentType) -> pd.DataFrame: