+++ entsoe_api/utils.py
@@ -1,4 +1,5 @@
 import logging
 
 LOGGER = logging.getLogger("entsoe-api")
+LOGGER.addHandler(logging.StreamHandler())
 LOGGER.setLevel(logging.DEBUG)